feat(athena): moving jwt logic to externa/ehr
Ref: #2115
